修改属性类型

+++ SPLSPOJO/src/main/java/com/zl/pojo/ReturnRecord.java
@@ -17,9 +17,9 @@ public class ReturnRecord implements Serializable {
     
 	private static final long serialVersionUID = 445088185055443253L;
 	//主键
-	private Integer id;
+	private Long id;
 	//债权信息主键
-    private Integer creditId;
+    private Long creditId;
     //回款金额
     private BigDecimal amount;
     //当前回款期限
@@ -29,7 +29,7 @@ public class ReturnRecord implements Serializable {
     @DateTimeFormat(pattern="yyyy-MM-dd HH-mm-ss")
     private Date createTime;
 
-    public ReturnRecord(Integer id, Integer creditId, BigDecimal amount, Integer term, Date createTime) {
+    public ReturnRecord(Long id, Long creditId, BigDecimal amount, Integer term, Date createTime) {
         this.id = id;
         this.creditId = creditId;
         this.amount = amount;
@@ -41,19 +41,19 @@ public class ReturnRecord implements Serializable {
         super();
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getCreditId() {
+    public Long getCreditId() {
         return creditId;
     }
 
-    public void setCreditId(Integer creditId) {
+    public void setCreditId(Long creditId) {
         this.creditId = creditId;
     }
